moved shape type classed to dedicated subpackage
