+++ src/gameElements/shapeTypes/IShape.java
@@ -1,34 +1,35 @@
-package gameElements;
+package gameElements.shapeTypes;
 
 import java.awt.Color;
 import java.awt.geom.Point2D;
 
-// import static utils.Constants.Directions.*;
+import gameElements.Shape;
+
 import static utils.Constants.GameConstants.*;
 
-public class SShape extends Shape {
+public class IShape extends Shape {
 
   private static final Point2D[] SHAPE = {
-      new Point2D.Double(0, 1),
-      new Point2D.Double(1, 1),
+      new Point2D.Double(0, 0),
       new Point2D.Double(1, 0),
       new Point2D.Double(2, 0),
+      new Point2D.Double(3, 0)
   };
-  private static final Point2D CENTER = new Point2D.Double(1, 1);
-  private static final Color COLOR = Color.GREEN;
+  private static final Point2D CENTER = new Point2D.Double(1.5, 0.5);
+  private static final Color COLOR = new Color(0, 255, 255); // cyan
 
-  // private int rotation = UP;
+  // private boolean horizontal = true;
 
-  public SShape(int renderSize, Point2D renderOrigin) {
+  public IShape(int renderSize, Point2D renderOrigin) {
     super(CENTER, SHAPE, COLOR, renderSize, renderOrigin);
 
     minX = 0;
-    maxX = 2;
+    maxX = 3;
     minY = 0;
-    maxY = 1;
+    maxY = 0;
 
-    // move to center top of the board;
-    move(BOARD_WIDTH / 2 - 1, 0);
+    // go to the center top of the board
+    move(BOARD_WIDTH / 2 - 2, 0);
   }
 
   @Override
@@ -42,7 +43,7 @@ public class SShape extends Shape {
   @Override
   public void rotate(double angle) {
     rotatePoints(angle);
+    // horizontal = !horizontal;
     calculateMinMaxCoords();
   }
-
 }
